+++ lina/__init__.py
@@ -1,9 +1,9 @@
-from . import math_module
+from . import llowfsc, llowfsc_sim, math_module
 from . import utils
 from .imshows import imshow1, imshow2, imshow3
 
 # from . import phase_retrieval 
-from . import efc, iefc, pwp, utils, bsefc, llowfsc
+from . import efc, iefc, pwp, utils, bsefc
 
 __version__ = '0.1.0'
